refactor: JsonViewer.tsx の機能を App.tsx に統合
- `src/feature/JsonViewer.tsx` を削除
- `src/App.tsx` 内に JSON 取得・表示ロジックを統合（MainContent コンポーネント）
- E2Eテストのセレクタを重複を避けるために修正
