+++ src/App.tsx
@@ -1,8 +1,37 @@
-import { Suspense } from 'react';
+import { Suspense, use } from 'react';
 import { OptionGroupToggleSidebar } from './components/blocks/OptionGroupToggleSidebar';
-import { JsonViewer } from './feature/JsonViewer';
 import { LoadingView } from './feature/LoadingView';
+import { ExROptionEditor } from './feature/ExROptionEditor';
+import { AuOptionEditor } from './feature/AuOptionEditor';
 import { useStore } from './useStore';
+import { getExrOptions, getAuOptions } from './logics/api';
+
+/**
+ * メインコンテンツコンポーネント
+ * Suspense 境界の内側でデータを取得・表示するために分離
+ */
+function MainContent() {
+  const selectedTab = useStore((state) => {
+    return state.selectedTab;
+  });
+
+  // React 19 の use() フックを使用してデータを取得
+  const exrData = use(getExrOptions());
+  const auData = use(getAuOptions());
+
+  return (
+    <section className="flex flex-col gap-4">
+      <h2 className="text-2xl font-bold">
+        {selectedTab === 'ExR' ? 'ExR Options' : 'Au Options'} JSON
+      </h2>
+      {selectedTab === 'ExR' ? (
+        <ExROptionEditor data={exrData} />
+      ) : (
+        <AuOptionEditor data={auData} />
+      )}
+    </section>
+  );
+}
 
 /**
  * メインアプリケーションコンポーネント
@@ -22,7 +51,7 @@ function App() {
         `}
       >
         <Suspense fallback={<LoadingView />}>
-          <JsonViewer />
+          <MainContent />
         </Suspense>
       </main>
     </div>
